+++ src/main/java/com/example/demo/repository/ProductsRecentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductsRecentRepository extends JpaRepository<ProductsRecent, Long> {
+public interface ProductsRecentRepository extends JpaRepository<ProductsRecent, Integer> {
 }
