+++ src/main/java/com/example/demo/repository/RatingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RatingRepository extends JpaRepository<Rating, Long> {
-    List<Rating> findByProdId(Long prodId);
-    List<Rating> findByUserId(Long userId);
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
+    List<Rating> findByProdId(Integer prodId);
+    List<Rating> findByUserId(Integer userId);
 }
